Rename and isolate getHolidays function
